+++ server/app/modules/example_data.py
@@ -141,16 +141,16 @@ class ExampleStocktakings:
 
 class ExampleItems:
     ITEM1 = FilterableDict(commons={'name': 'Spray', 'vendor': ExampleVendors.VENDOR1.get(),
-                                    'warning_quantity': 4.0, 'unit': ExampleUnits.UNIT2.get()},
+                                    'warning_quantity': 4.0, 'unit': ExampleUnits.UNIT2.get(), 'purchase_price': 60.4},
                            setters={'article_number': 'sk132465'},
                            getters={'id': 1, 'article_number': 'SK132465', 'quantity': 0.0})
     ITEM2 = FilterableDict(commons={'name': 'Pipe', 'vendor': ExampleVendors.VENDOR2.get(),
                                     'article_number': 'FO213546',
                                     'unit': ExampleUnits.UNIT1.get()},
-                           getters={'id': 2, 'quantity': 0.0, 'warning_quantity': 0.0})
+                           getters={'id': 2, 'quantity': 0.0, 'warning_quantity': 0.0, 'purchase_price': 0.0})
     ITEM3 = FilterableDict(commons={'name': 'Screw', 'vendor': ExampleVendors.VENDOR2.get(),
                                     'article_number': 'BA45678', 'warning_quantity': 3.14,
-                                    'unit': ExampleUnits.UNIT2.get()},
+                                    'unit': ExampleUnits.UNIT2.get(), 'purchase_price': 65.5},
                            getters={'id': 3, 'quantity': 0.0})
 
 
@@ -231,7 +231,7 @@ class ExampleWorkItems:
 
 class ExampleConfigs:
     CONFIG1 = FilterableDict(getters={'app_name': 'storekeeper', 'app_title': 'StoreKeeper', 'forced_language': None,
-                                      'debug': False})
+                                      'currency': '$', 'debug': False})
 
 
 class ExampleUserConfigs: