+++ server/utils/fill_up.py
@@ -111,7 +111,8 @@ def iterate_items(data: dict):
             article_number='{}{!s}'.format((''.join(random.sample(letters, 2))).upper(), random.randint(1000, 999999)),
             quantity=round(random.uniform(0, 10000), 2),
             warning_quantity=round(random.uniform(0, 100), 2),
-            unit_id=random.randrange(unit_count) + 1
+            unit_id=random.randrange(unit_count) + 1,
+            purchase_price=round(random.uniform(0, 1000), 2),
         )